+++ src/analyzer/expr/call/new_analyzer.rs
@@ -232,7 +232,7 @@ fn analyze_atomic(
         "ReflectionClass" | "ReflectionTypeAlias" => {
             analysis_data
                 .expr_effects
-                .insert((pos.start_offset(), pos.end_offset()), EFFECT_WRITE_GLOBALS);
+                .insert((pos.start_offset() as u32, pos.end_offset() as u32), EFFECT_WRITE_GLOBALS);
         }
         _ => {}
     }
