+++ src/analyzer/expr/unop_analyzer.rs
@@ -32,10 +32,10 @@ pub(crate) fn analyze(
     }
 
     analysis_data.expr_effects.insert(
-        (pos.start_offset(), pos.end_offset()),
+        (pos.start_offset() as u32, pos.end_offset() as u32),
         *analysis_data
             .expr_effects
-            .get(&(expr.1.pos().start_offset(), expr.1.pos().end_offset()))
+            .get(&(expr.1.pos().start_offset() as u32, expr.1.pos().end_offset() as u32))
             .unwrap_or(&0),
     );
